requirements.txt
langchain
langgraph
langsmith
psycopg2-binary
sqlalchemy
thefuzz[speedup]
pydantic
pandas
openpyxl
python-dotenv
